1차 추상
